--- electron-builder.yml
@@ -1,47 +1,0 @@
-appId: io.settlenet.xpubviewer
-productName: XpubViewer
-copyright: "Copyright © 2021 Crypto Garage, Inc. All Rights Reserved."
-
-directories:
-  output: dist
-  buildResources: assets
-
-files:
-  - filter:
-      - ./main.js
-      - ./renderer.js
-      - ./preload.js
-      - ./index.html
-
-extraMetadata:
-  main: main.js
-
-compression: normal
-
-afterSign: "dev_script/package/notarize/notarize.js"
-
-mac:
-  target:
-    - zip
-    - dmg
-  hardenedRuntime: true
-  gatekeeperAssess: false
-  entitlements: "dev_script/package/notarize/entitlements.mac.plist"
-  entitlementsInherit: "dev_script/package/notarize/entitlements.mac.plist"
-
-win:
-  target:
-    - zip
-    - nsis
-
-nsis:
-  createDesktopShortcut: true
-  createStartMenuShortcut: true
-
-publish:
-  provider: "github"
-  releaseType: "release"
-  vPrefixedTagName: true
-  owner: "cryptogarageinc"
-
-extends: null
