+++ package.json
@@ -1,24 +1,19 @@
 {
   "name": "ledger-liquid-xpubkey-viewer",
-  "version": "1.2.0",
-  "description": "Electron xpubkey viewer application",
-  "main": "main.js",
+  "version": "2.0.0",
+  "description": "Ledger liquid xpubkey viewer on webpack",
   "scripts": {
-    "start": "electron .",
-    "postinstall": "electron-builder install-app-deps",
-    "electron_build": "node ./dev_script/build.js",
-    "release": "yarn electron-builder --config electron-builder.yml",
-    "appVersion": "node ./dev_script/appVersion.js"
-  },
-  "repository": "https://github.com/cryptogarageinc/ledger-liquid-xpubkey-viewer",
-  "author": "GitHub",
-  "license": "CC0-1.0",
-  "devDependencies": {
-    "electron": "^14.0.0",
-    "electron-builder": "^22.11.7",
-    "electron-notarize": "^1.0.0"
+    "build": "webpack",
+    "watch": "webpack --watch"
   },
   "dependencies": {
-    "ledger-liquid-lib-simple": "git+https://github.com/cryptogarageinc/ledger-liquid-lib-simple.git#v0.1.3"
+    "buffer": "^6.0.3",
+    "ledger-liquid-lib-web": "git+https://github.com/cryptogarageinc/ledger-liquid-lib-web.git#semver:~0.2.0"
+  },
+  "devDependencies": {
+    "copy-webpack-plugin": "^11.0.0",
+    "webpack": "^5.75.0",
+    "webpack-cli": "^4.10.0",
+    "write-file-webpack-plugin": "^4.5.1"
   }
 }
