+++ src/index.js
@@ -1,11 +1,9 @@
-// Modules to control application life and create native browser window
-const LedgerLib = require("ledger-liquid-lib-simple");
+// import TransportWebUSB from "@ledgerhq/hw-transport-webusb";
+import {ApplicationType, LedgerLiquidWrapper, NetworkType, GetSignatureState} from "ledger-liquid-lib-web";
 
-const { app, BrowserWindow, ipcMain } = require("electron");
-
-const path = require('path');
-const url = require('url');
-
+const testingAuthPubkey = '04b85b0e5f5b41f1a95bbf9a83edd95c741223c6d9dc5fe607de18f015684ff56ec359705fcf9bbeb1620fb458e15e3d99f23c6f5df5e91e016686371a65b16f0c';
+const productAuthPubkey = '04fb353afec6e71290869e7146eb18be90c25a1d90d85ab6dc9870d29d41da2dfacc553e008621c3e78f6dc78fe1b3f5c0c200f67c8dc2f3536f1ea63de485dd88';
+const appNameTitle = 'ConnectApp';
 
 const proposalTx = '02000000010354cc5be98996fa6e7128922c65425652746d2b9fc57576e9c54cf73f9e245bb60400008000ffffffffacc35ae4c07c0d73110c97f4b6d7fef5464f3c98dd60b8b2f655c2e867fd60f1b1a2d523532697e23070d333db069339b3ae94a15a075ac7f8e4f25ce86d7892081a68eb432a7fb0f286bee1c2348fa2d84353b2d3b8350c1f518958a87e21ebd30054cc5be98996fa6e7128922c65425652746d2b9fc57576e9c54cf73f9e245bb60000000000ffffffff54cc5be98996fa6e7128922c65425652746d2b9fc57576e9c54cf73f9e245bb60100000000ffffffff050ad6f88035d77c23ae409242de725e8e10a889dac1c2b2a9b06354ae426e31d0b608fe2e688f7c5005d64cbb0639f9eefc79ad7749a243317c8b9db747bb94f50ec002862fe34b4b6607bfa362065965ff0e8da5936f5a4c640c49d0578a879746486117a9145147295169feb02e6de34f2443210318de96d146870b80b107e44529efaadec58dae4af49821bf943a1ff8eb6f9936ecd3fa6c8e93b0092f260068d35adb85e242270ef89d4921bcb09aede102d4a4da3cd4cc63c4e7dc03529f1d5af508318baea0c72cad85a78bd4f5c94f9be480de99e4942779a95ef617a914c89964d7512214d8cd4cd8a2c2c57e7f4238ce3e870ab06a2ef6f83003ceca11d177d1fcf6cedf57a166ff4da71bd97e6b10768e8b9908d415daecdfba9228eee863ad881b3448760d9e55d1475aae0e6517675362e86603af1b03fd752bbc05aa60fea1a235b368e5ffc9948955e23cb72f8f76a79e778a17a914cb25143d64243ea88502dbf8d761b7765d19f977870125b251070e29ca19043cf33ccd7324e2ddab03ecc4ae0b5e77c4fc0e5cf6c95a01000000000000c35000000b3aaa8332392f4bfad1ffbfc1c00ef77999892efb089376c9bf31e3d444ef6fab09c125644453422d3411630a44405e0c67f1ef441dc284386cbea79bb91229686a0385f7b6c6de2bdaa307aae7a0085f80503b70d2cc01828910fa5cf359aaef6a5517a914f6d243bb579d8823d945b8415bf59c76400fd3658700000000fd4610403328cebf01bac61c8768beaf125ce55137cad91e320c58e3adaa284e4bba381c2934a428d517a99af681936c3634160b53b2b8b032fc102279fe9eccbe47c46c936424b89f0d2bcb096383620a254cf616d4d72679c1753b75b32bd35618a81c8c614dfe5977e511eb67bcd28f5c8bb651362ff982008705e90e191364d73d87b1da1a0cef4644a14c05001f8eb87a4f983a87a50777e0e4f11326dd1d49b8a8bb9042275dda9909a5c20ae37e2f70c5d6fbc1bf8f298212bef5ebd642708736c09a767f48e704ade2820e958b0b96af2390327c95db25f8e53f77e2260eec551e3865fba51e06259fdb2ec9ebe9dd60c982181f0670a073ade76ae86ba5750b0696ba61a0d06559c4fdc7f2270d7dfe7daeab07426175536ed64bfd9550289f7445aa2390ee58e9f0867201aca40c9d3f5b92a90eb5db4587b0c71e864e9392268a001e5bb18cb307c3c7b94e5ff15fe094fd112581472b7211ab979528242bc893975d86179a4954970a16f954baff0ef63970a0e1896348ce758de3c62e86811929f4fed2ba10df7b7769aecd45edfcbe50f57456e0da356672ca8f8b3580730261fc2c14123552c85714eaf12a3c93246de3c6eaa8214210759679f60322c24d94ff59667500be2abba8ba45038e0b32c231e00ad59116895d22a5a25fee2adf2bd48139ad5daddf032315a3e3d10515bb5bf6dc05bbaf9b08b1250423725cc7dd462b9d66bcc8496f9aad8b4c0a5565cb670963d716afafbe37f06b6be4d3d3569e206642f92a9e448f2087b82771a9920076a194363fc198456d7f073cb692d122f034d50c945064f20bca6f606c61a951638a2721471014edfa0c70d8b6ae23c13a7a5748c87bb75cb55344b6e0b06760b9e7843b055dafcfacade24dcdc62126a66082189d2a81f5449a33cd4c1fd621f3048aed28992a6dcc00a54e4f36c9cc0bf9575b1d2abf3e3a3c6ef970bdf634be39d2d765caaf9cccfb8827d26ae79a9a360f2cc3dfd020acc331b23dda5688dd60b69fecbccfd8d75835ad3aa8e66ee7015edf28bb57088549271f26ef24d4991e7ba3873590252bf23d1bd5315f23038d327ac0069697b91153f2225c60a6ed6b76b5e1956ba7450cece29dadc6329036f2ade41b5753b5afe0c967818d4a9f1c7c10a88e289db1996221e932e2e72a59fe4b5ec7f2af0d67887fb0a836508628e50ea7309bc13809141d59de26998836952cd5fb233a23f02abfa834bba926604b836707bd194e0aa76ae391c0ec8d1656b042e713e2b3b8584711ad219f3a682139a79c9d5cec32883972157c8e57748d367f4494ccc62871f39b5cb080a994e7b624f7b494e0e46c9b52ef39603a6fffd74129270e67eab6afa526df91cb5b6b8fdcbadd462923414ca003ab5b84b9d0633bd52c0cc7cf5cebede422ab481d85de75028b0873893cb747831e02f7c8e14e46dd3d49fff0a83f5dd16b254c5b68a01944538afab4fec21641963c350f79336be57ffa67664164e97b0f86849115f022820fe5e7f94911cd76596ed9d80cca25f30ca0ffc270d346be2d72521837b2cf5681244c5e0ac9d4165e50600f09d9ee066f40b1a25fdd93bfa2c62753903af5fbc240cee0f40eb84a8d583a46d4aa455e175d1450f082051ea3bd725c3ca8dd53c071e58c1f0ea3c29dc77a4ac3cb7b9116354a2e3c042e2968de826b757b132b5c84fd6e6bd2a4102927c7899301d1ed58a6f23517feb2fa4aaf5799cab3d61e6d3c36507b978771525e7ce8df4906ee47e1a70ac8dd5c076d9bf0a8c7b28ec6baf57048c8cb0bf8b98b531c6d39d66f324bb0a4817e684ef637b7f2458768ebcc5db012573ce3ec1beb49a9047d22cbe4749fd01fe0e9225b90d8bce7be09c12aa6f11addb67922ff51a72572dd90e739048a836bce388d397c5ca8ee00f4ae3a656211b183273c0356e05b5f30f241326d96044a8d2d9c413c75bdfb346d8e677930fa1ca90e966afa0f62d0a223e93b129ecf03a169cc2427ef22e96393b98c5eba39b7d7ca2d0cbe99599ee1c2ed5256b6e50d7eba8051f3622b837e55ef1c107ec0d610ff9498f73e93815226a6f4ac29c55f4bee906942b25797654d9144223f3da6cb04b96a2e67261c8d15e8a4990fc68026e35dc5205f474c15494c84ec59f4ebe5d11c11617410a65e84fdc0334c8eb36ced05e80e5992da59fea5cd860fc7a9c156eb5e6d1fc407e12d71566a13dcc53c8511bc2cdc61e4ff150fff6844a211df62f19fe67478c67175ea2ae49f7add986474ca67de9e46f8d75ca8f9146a17be7e2e40b9f73f524d7e811c01e158185b8ee380fc41270132b83276087edf5518479685fe89fa3e91bee4808ec4ff3a342fb7ceeb9628594fd610a26d8bd04ffd4fc06d89fbd3aad7b52c648cd02061d201f4ecf7717a68cad28a71da817e2b2b2d4a831d2a58ff9ecf0289fbcd0414a7fe4b4d1b11572731f22e2a4948b46b664a9d86fb1192d8267e48371df699701d061d463a2b73de36ae6b3f887fca698c3c067658980dd252ff4733be9e5aea5b7fb06dadb928b0fd9ffb8de58fb8dc8fba638e2193c0fcc58aa1a3d4d5c85bdcc0257596255034fa6fa1a6ed88e9dda75bff3935bd6e7aa11b32e7a6bb8ce728d577dbec61683ff5c5df7fea38275cd0b605263d493c3d1aa9f23882d46107d6439d654696a97521d7dc874ae717d956b50867f725eedec982184531e8596038fe9441ebbd5d3947465c8baa5b1bd9c1db6ab74d39e93f41317b0a4b7cdec46ca992f5a9bb4f26aa4c6034e4d729f8d65cd6d4357a35ee6fd18fd75c58d5a76787452caba2fc593792642cf5c589bcfadf80aebc44423146e3924065ed88a9966005628e24f5d55aa0e78b118674e24274789e30f90103c234d6c810d62be56f151ff1e5809b4f9e1ecdac13399dbd4cce1bcb9e8dfd5a2fe79ff7720ed03b21b6e54b26667543033a796b274a0a54e0b6af5e3b4a30e336d11bbb670df998cac780123788c73380c1c37806be2f9daf7dcac75a35449648490b14a06510a0d8c484424240edae720e2e19695f7357e2e079384caf3a2fdae0cfdc27eeb075d2d8e11d8fe7c0cc441aa6408fd00d9447ac17e0f2dff6491d3787e8b3f4d672fcb8886b4a8c34495767e422cc5988c07395ff472cb7f757e52138d75995c939411dd57c78b40193d03a3a15ac61b6b851e955efb34f1f0116b49ca010032ece14c3f0cc51e84d12ef854214c8a5ce26a8d939df17f7c6aac15f8268e0b700a11e16a4853fc8ad0298d1d3a3acae4195cc96c5f7298c0aca0b506ebd0c9f52b4aa557a55872371e782d1d45ca3acc3fceb5f7272da6fe4d76c9e93b98b60a12f88dda43c9055cf61132feaf91206f9da3c38bd6fa55ab26f146cac8e9eff84933aa897ed3797067af7470d2473e1fc15a0bdf816572900de1674e42b5f41fc968dd921b54b9b18adb1200c595040b05386d487df7b8fbe4646507dd7120bdf4083fc2e78c79f2f101697e2f8968ae454d07cdb7e496f7fb0e7a210ae20c004c500421200dde1f6b759907552b5d1b5d9375d9fb6f4e84a1ea834bcf4337d4d693520ed7a6eddac5bf0ca9f87bd1a7731be1bb28ec8283fdb26ba56e223c9c6dd904f3ea9c68d05e2a9d94abdbcbd11d95565272aa694c24350cdb1a261cd31519325a1601d2c0d716a2060cdb4d85ba69e2eb907e876b078c01dc1067256d02282eb1bc6d284093284cd50a8a7afef595106d0f9b91a1a0f86909fa4396f8b5b0591894ebc00b77396d78a1ad3d7971050427854a21d3a5e0c3afec2df83c46715538d6c5b157edbeb8863ca5e0249debb9437edc7a03a02d2b4db883dad9e1d3c2927b32500f70e6a7e81276e84a5ee72a4b174bcb74b30598bf0bd35a6d358b665ea3a6ad066e57f5bd733157e48d0677f2edeabce643fb87c7ce076bdac6af192b8c570ba1bea972fd032533bb3e908d9d910daa4f1c8d56d5809472bdc536b16d25f0240dfa1193ea333f4b911cc2287144e3cb84559d215686c032342f0b0ea7502fe80c13777aebc3cb0ca7f0883dc002749cfa196e52c339db4156ba81bdb507c000c4761cbc48ab725e7889ea64b3c53b7c5f4fe14bbf4219cf5f605299a606090bb1f2480131a85944ff8a19366ac81956db2f64d30fa8a373fb5428fb1dada3f4f33f1f266e76e56ccc49f56aac801b1499f0c5ae1141e8837e277f1c3954c244756798d1f3f5ac6b3984d7715a863df3079b90604f30aee0d25dd8f5f3492ff5cfacab3d1a56c4e6b12e9aaa51d878cf1b17308321508702c44c2f875efcb1c9720f106555d032ef715daa5bc0a940642566c6f1403eda4531cdb64ee9602d4b94ce05d22bbb63b1e49cbf25ba3b7819f0b89c61b7119bd54dec6a39dc418e69d82b31af65d0cd047b6a44092ffe2777d11b4a489ae3147493903810ff9a4f676db3861adb280a215187fdc9dd4ad7391cd84cd30608259e396e7e105691168e8c9ecfde00809d4048141362e39da874ac3c6dace9cc09e3b8a4a73da47017599df0c11cf0563980df259d04a1a6a68df0278074da62239bddf0327f93233e12971d56755e23be57c893e9ec293b402c86331c267484ac63dff794e83c9d69fd0e06f78bfc72948684925ea336236e99716c6749516003ed4976f84806e7c8006d67ce21a4e8ce8a0b6d204ebdb5cba72747cd4c3d6d42df958093385686e1b5feabdbc5d468f55daaa16c78e2720d5616b8af22f786b1612b43af8ded063601b60662bd3430bc7c8396dee51a7b1d5839c2a231100c123f595810829472f476b1174350ba13fcc0a45efd926e2bf5ee8b9f50f597d9c7b6c3fccd7098f33c418087a1b3a3a61a5cceb8d8cdcdf4f9d02cd2f7660972b0440ea04a4038e7d15fbb49a41d79e580de6173b8860d1a37c60e4c22a3a7f1b1c013aa5b4401abd3d8415c15bc59467ddb6fc6fd86d5743b266a06a91e8f7e59f6c68a8454c8931d34d0c42ff2698f3ada93a352b97f143ffaa19ed07ea2af5ef972c62279c57a2b63e87aa5934e23d6b6deb1191fd2a78aa22d6e3fb678da07b83d70b687a5eb7cb02013b658e4b69209890a44c45635e7020030a7dc080a30bfc277971279c438fc4f2b1634077eb2993b5a337ffc5619d401b553acc261db7b0368e8804cad7a46c9aa86f233dd2e4d6e9a7df1a9a76b01f27577ddff2cede7e16d292967bc5c9853885a05bfb7f80368cfdeec50c32fd33da57077fd12a813f8756a1b2b1bc2c5fbc21b61aa7d5685caef35d9259ea06973f429748f08578d09984c3693bdf205720c3d363fb05b755d58640d6cfc0008eb0928bce598e5918fd28294881819b53b6ff12af7a47f9f39df8d3b15c3a54088c2eb92720a698a08679b50ac062b1710c062f6c501d9a864f50b96b796c1ddb008ce15f09f7a74d0dfaa2fa3bf95e4acd8dc4bb488055f4bc12f9589c273f6057277f7fee43d3f10435e13e229b4cdd684d970f626739ca038aa7a74e69e768569d252a10db5896881c54a4678c78ab6c788277b5d1f6cdab11d8ed0250eb0620c415a53ede8e414267acd931b6774822daa156f210aa701c9c4dde9739f8479b84af44f471a934bb44d9eb071f6ae1deb377e5cb9ed28f326dd41bda9b3dae7b3e3e65f85a3a2ceb0324aa86606a268b124eb1b9dd5869a5eac7b142d6162a30c2653f87484b68449ba34cfd2d35d37996eed3d65bbbff5485b56cee74b05c2771b7e6b312ecf6c83568c20cf30fe1d0a718ba330f9a2de8160161f7ae97ae100e2a4710e4ffb7f702d12bfab852ef68f60bcea8f7da6abf27912b835e89dccb95a8f383fa0fe47899b110f36f2180c606f5eaa3df7b371da253d11a98f6fb00000000000000000000008304000733e12ac9c2761e92aaf8cd0818e871e4bac86546ee92102f927f30f5cb73d4d808dcc021963d841a601fe61a71a089e9baba699a6274cda2d5fd7f1a00010528a16ada0969a4df5016aff8832b1b678386f51ed3e4757caf8d0f233e4dcbe89b9ab8dcdb9188742c7a6a141c4282f1e0894012cad5c11f75aac073a314e17cc3fd4e1060330000000000000001696aa60122ff3bac63b6d0c00308ac85fd43c76039537b1c3e41f21964b46a917168261222883fd7f26642479921bdc40a27ae183380da7172035ec2a06f76275bed497318a0ba3865ddb02ac5cfe89d5e72ffe026aad9ff5b15a4bf95eb885504f3d28c48da2434bf4f506ac4e31f90599f873c66611dbee72dc0d210734894006519c7eb45a57cb7ca1d68520fb653396d885bb77957cb40170a030c21a03dc37d8cf9796c3f47384a31a911a2f2be347829188695c05e641ffbe89608865b0e239b480680d8ef6579baaa7c948ca338045982e695c0a2b95d574fce5d99f5309a39827373d0b43e2e51b5750a1b84a2f9576589fa0b56bcbd62088a99efb76f3a4d02d904dddf527c284fd1ca309ea9438cb3f38e3e1641771dbd086416012463b09f9cee9f2fd4af6c28bcdfed6be43ec197a89d22057b6efef44920bddb3d5052c4ca1e1687d3a9100a26a23583a4306adb9a0514373f138aec06e423695848ee88b31cd9a9c6257124e974914760146bae941d816dfc4128dfd6f600cdfd7e34e1417068e7a52d720147941aa52021fb6c71dfea28086ae9f11bd2c47c4137dac44a9913811320de45388fc8af63612068f9edb3b9207f79034f79a5372cb32c4224029b701230b468902322296562f8b60abf574f4ec11843533d6014f927ec9efc606a6d4e9a9c4e4b5403aa0c5c15c3b0ebfdd1baf59e9143d9ba6c17097fea3f03bb2e377eae45b96d6e0bdfc705e9cbc7dfadf4e53dd5b69b7c04f9022b8fe19b562bed08e59d64500db6606155cf317ef4904778f6e1a5e1a7c3a93e4dd74db5da334c03261247d5ec9b80ed38b67448e3ced60925ee251056b17747b225282784e6dbb716811d08a198fd4aa4f1f1043d4553feea9c58d531eac735c803c0846ef566244ffb35546ef80b185002425dac7c51f17abc8829d498cbedafe67f0e1be2392667bf4cfaa067d925b0633ac995b83f7d5f267ec800dd6cc91e5601dec23a104e4dcbe1b50b8f4024d5e6aed77eaa35a835d8d5758de8d8a819894b26cfd3d01c8dd5499fa40a5561b1c98df51d481965c1d33e9d76d2b8e5c39f8f947ee3e55ba8ad81c4e588af15e497303da5aa7ce73ca75e50541cc1a679dab32ff63345d566015cfae6ef4e6e69d737e3b0ff60d86ac667bff2c14c2bc85e0480f079a82ec463b6c6e5cd8d70e9a7d0dc48b5e02a646762361a8b01381089773f1b926308c9fd297573c6f92f9db23ae426ca48296cf39988cdcacca16d9e3952c88283d7446daf8664aa184b5c5aaa08ff2f9b1277972d4b7585ddb86a20566d37558e5cc561ab2627773ec86dd6e590f73354e031433810d37ffc4b6c644a5cd07eb96e84918009c79bc9597fd7b5a70fdae09066023e622712f78201182b0b312cbd149f1e0cdea866b52e6563a5aa01934a9adde430a05df4dfddfc06d994b11c9be3758c17f397e65c825c38c5bff9453aa29b5533918c2a4fc2364a116b489d89e52aadd338d0871bbd713f42788693c97a1e45a0fa0e0e1365ca9f8ac01822cf581e9b987dbbbd91d77b2cca4c64e32d1c4b8d6a1cda690d07cc8dab5f5cb6920b461509c3166c237eab176f9c922809dddf38714e3ddbf656a0805cee88a03967fea1317a70d391f1414c0bd178b53320b51a602a29909a387894d2809ac76ee77441b04afa16532a85a87cee61abf0b1a74e894ca26e984d09ed2b3d0499a619968b081c392635f553f7e1ef60d217684cc5ecde70a7a3592b18d1a6fd89f2a6d06d788fc200efadff3a27e1d2d589ae78243570ccce495de84cb8baa66c2f93de991bba46e32e4aa9d350d3a5341b2658841f828eebc925ba2a7789f8a7fba61bf47e7a230e2fccbb5ce586dac12a139d8b98dd93c59a0e4dc0463b5c5f066f07ccd04790448986889da78d85d5b3b822e1009708afb9e1cb40979af74c8400ce61dd7099560a35ea09e234f905ddde59aac5c238526e17172cb6b390a7aab9468dec7889d440bdd1005769cfb5557ff25e71dd1a9bff38ea1783cabd13bcabf7a96b7b54914ebd56c4f32afe08a8a9b583a3c90328d666d3d3775a8f90a17c02e197b807aed65b3dc63101758fcaf7fac2797dc18937fa744026b7595b326e6e221530924d785d8115721c7a2c56bb01eaad303a1f6764fc918b4c2bfe2e70a78dfca8194619f8a9e156acfb878783c4f9823dcf14a537e80a1771161cefbbf17c88bf495ab8c0de77817a99375cac265b0468415713acbf83a956e04e694ab56f220da3fa89afb63ba3c5d00b310e4419ed469f945de0f2ce5f04476fd2719cf5483052f73dec3556031f62a79253f5379b20271b42423cf6a59d9a7e86bc99f814306697066fdd47551b75be098de040f89a890a9bf7b302c48b7fb4e76cb42959f37a6fb883ee94dda3fd5c7c689c5d70ac4938f0a1621a1025f43d0418efc452eaf682923ad374a88d68d8bdd1d52a9b61754a35b014ac6dca73541476992a01201ecef86fe449ca9b185b8689a67cecb92c3051e5b0d3565ad52ff9b8bfd6f20404c1af7e8a05b50852e91365167895bd9124af6bd6495b93ec65cc31905db17904620c1c4085baefdb148100143087466d65bf0796d1fc8854cb028c8a7463aa1d87294cb2d6db611789d1c19182e2ef4905e38f0a2414a9a519f05468575d5c515654b14a860ec0f649e3ff2cdcae5b5b51318c1189c7fd887b606c0cf37259d0ebc30c1d5761bf4549e5ef64065404908ecc615cf51e7d9eb42c5b10a6d71675c8f5da6fb1d5a4625c22b55105586606b7d018b8caa90a0716bb0313992e7df9eb68d52e206bca307e34cd3135e3785b6a629e15fb0d8e1f2c913082dbd52b223d487aa2c8f66a33360b7694a5a374d51ab768533a878b39b65ec8b38506b96d53bcfb778ab9489edf2cdb3e2c6f5a3c1e14f6021f80a460ce73b8d70f1c5aec2937529980cb5ed753e9e79d5ca3bdf90b1a9380dd42a9765b6182d5a07f7589dc1041fa4aa37b25240615a73a7fd3c8342990d597e839a8a8f620a55a10a15c4de81f7cb758bfd94eadcaf06b938f92c986908bdd33883a002f2bf7a0e5fd76b596c092c89c070017644b813f854b7fed6ae6c5f6d25ee4643ab8505e2339f096a55836d93703ddc1a5d65c9388153540c51366692dc767b81bbdf0f403cccfbf324622cede2cc62e7b401dafa740a816790a89df5c1b7a1cc5f1f55b1455ab1809ed55ea9337fb9e64630e7bce22eb9b0365770563fcfdf1f84937772eb1e8307850fdfff3765085ce7951b98553081035be2db51353f85af34b82904bb85c1c95baadf2a9f3efb21df9918c9a779b9591d4812e24513813ed382d31f193e8a228999eb23aa6c8760224440633d9dcdc9449d6b83ee31e3a921b2b129abf93ca775064dde53f93cdb50343d239183133ccfa55ca06dfcbd78d2a424bc1154e868668fb8960e4ce9f6657f8f458ab0c3268706f8460f15237554ef028c4c9a616091301694721ea8403c5817e7bc736f1c61cbd2a80d0becc9cefc0244965a7500706b7e05288717823a06b80794bfc95ab5def7bf8a28e634300b6d94461d489be5ce08a63b4067930538653fef07519780051899d8c9d2cb790a1b2378c23c7d153210f57668e8be6e2a410d1a2419694183b71dcc4a803192354d8b20e7263f60c0fd49a1b33c800dc6f2056b447bd99dc43d7b471ef6ab9e5c1bcbb5dbeb01d6ab64a55404f4823b613f648c11056781561a1d90516a234f8df92580d7594e329ed5cb912cdb1cb3a9f65fe17eaf8c4f26321bd8bb14e8a3667ad41e9c2f4b4e440a1ba91027af4dc8565b18c017ee4f10d443120da838e1c5ec72a6175cd057c94af3d94c066c5fccd4a53e1f2801bf655fbbf12842f02f4344bf377b099e841fea8c472f3de25736368a990414180fc2a5febcf002e0073823432a5039be68975a25465f69d9572dabc833c5e5485651839cd7ac71fdb887ddf68e8dba1155ec6687aeb89a5438408c610d7d11a238d7aa55335ea6a018afc7b1f663d7bfd2ad5544ab7d5d5da110b51992441fff5ff41a8477491927c64bd3c80f0c31f06ab566e2804e348c861d6c2685b098e6cc37fcbfedb705f567f4c9e5a9a0626a4dcdcf2589b64edb551a6d589e01d01fc95e06090bc983fe1e2e9bd6dfeb6e487a1a8ea69fc614740e2480c74ebf4fa38e921aa6b2cdc58eccf5352ced192c2bea4fe17562132174bd7795d99b8bde487f7ac37715701b5a09172400bd2e41caa889a1e4d727e6b932b7a5c648546f735893d89ba11244973e32d58c2c8b7e05ce4f4fefd0ee6c8b5657ecaa6e02f1315bd2997259bfa001499d148a3d50c342eebee694ee4ab7ce296bc18c69ce78673e97b40b1b2da5b9ff0ea972368cc84513fd05ea856c8ad85818b7a7a552b86e2cbfacaeb7faf85a10930897f85d7e904ac05f2e8aabf68764f59e83da72688ee845800073eac72712000f66986ea7b4cf548a6b6e22e211e4b7768682345bae63fe857b19bf1b264dcb30e16cdc4d54783036b1e9605e841105f43595416b9f5a422abd45e537f15fbebd3b01bf9c72191bebf760ed252dd4c24daf2eaa02d93a8c3a3033842fabe746926e15856d58f1cab1b83439394eef3e74a18b6088622f41116caf8c3db51b02b5dc50f92915b215ac48529857ccfce468c0d003a12e231afaee79d84b01502f6dbe9dc58942e42c3806ce7b0c4db50cbf07aba375f251a1da3e5eec6b56dd54f919eca921b9b7ae37e2a9a3c41a83d486498e75cd8883cb1c3f698351c524896bd2790cc878be580982cebf92964d023134701f3241b1a169e1dcd4d72cabd46ea868e7ee9510ad9a0b4cc9d71a9f89b4e772489b8a9ca873c475cce30e7720636dd30b51ff852b93c4d0bbece8a994ecec7d601a1d6f71be2631a680f980327606cc85c19891d6fc1a9b1f3bced0b8be18bba62a5d9f936a6fd4db703959f543a1cde8f271ed6ad074d5771090cb2a2222bd107d8890549d33cdcaafe8676897ddab5687c250726690bd121afd41e52191ecadde5a2f263fb8c49144eef3ea044a3910fdd3710ebb777aa6ccd706de7b70e70076ae4d12e3ea0b68b40c77b7a7a0584b3e82f881a3bc3dd37ddd8dccd0ecb0e44ef48aa66d5a50ed95f7b3496ce34bc274a05e3a28df9d27f3b27742c1455ca1fb6704d8912357f0f9471111eb27c20db53e6a504e25d9af757211049ce25ab286e932d468b3f73d238478f8e3bd6cbde9cc3807023b7b83dfc833a761220b689979f6fe77682c6a994cc2d0143e1d07ffbd9426cc1d312a695e2a399621805675f4dcc3ace15e10c0015f3360afb71bb06d37bb2041cfebe116bb838c4dd5ab38824038f39af020046cefca2573cf297708271fc599deb48a53196491278641fb101d955fda6910f31c6873d269daf80e23f61e6a243b120461393d6f70a5be634399700c936adbb3b930077e194c6dfdf4ed6f653bbb64255a8fa47215a798f03f8aae9910423b019038b6b18848fb8f12070c72c081ae290e15c0a12016a4930e0762ae12bfa0f8864f707184b0a51069f0877c2e46636148b32d0edbdfb51e11b6dc770cf3b4d7ee95f780d7220490e3fb5855d42ea0286813a63fe9f97907ba8a73a19a9c8f3dffbf6f1de395361372cdf1b2acb6538054b46c4fbdbd07d92d3a52279b04f359b917b92e40f3c1e3714b43e1109902027235f503eb349b8c23a787f7989a08933fabafd4d870ad61e7da4fe90eaaa944d9f55e52e4184d0e79a271a956ee24aa0797829014256242ef113a9c213aef404e74c30a3997ff7ad6b6583066272844ffe5b422d982f79b4a9828dc5de3c2faff30d813c3fe464767e03f5d68304000e6a09de5ae7dd7c22ef82c7b7c946ae6830d00f8441188f81f9a4000aa290f2af61335768494c43ff4b443c7b32f914c7c5a2fa07cae2967751a59aa21101ad85bd83fae9dd1e45da6cf8ff5d021e6abef0295431dfeaf076e29b7546c8ae7cd225b3b92da38e505341b4737d4f1d9d513e69b61b2e7d2d14889109613a322285fd4e1060330000000000000001f2c748019d4f3c00569754c8f661e935dcc5d35a686187939fcc8e0a744bc91256e9e0f0a167615d1d1cf7a38aec618d5f09f4e0509b7e1963fe98a10ff315612ad60bd333f88436bc8b5f45439515316cd5dcb375277ba790477ddeb7e60844526d2ba83b1ef3e890ba1ab39a42b3bb33daa102d84aaab99fcd9574ea06c6bae2669107e0d526b79550b6f5d17d66cc0151c22c882245102d24ecebeedca2644ad2483d56285dc9bd29fd0caf369bda744edacf0dfb12116aaa21997e62c501d00c9beb26ebe6399719a0a618a02177ddb244171fc7f1d91b9a4ea14016e12a451c722f28d0abd88dbd6b63a05893f6e9417ca3e61751c6142be3d2f80f33bd640be8f8a7df907797e72cf00efd42aff6547fdcdc9141f0141c66fba670476e7cec6ce53bac46f46f978261f627a49dddf50a6c43b91e97536a3862ffd665be35e317b6f204acfdd83a6d01c0297af336b04024dd70536d6a1e9dcc6decb381a9ef91e25f6996e7689f43b3ab81d8ac2f2caa247663d8c58acbd3cc5354534349198b20eeaa38e11deec096b775e3519343a40519de7977288a266e2d478bf44ac0ef686192b6b57705e0aa20e81c7a687d6b4cba32fe4fb998c007c31f5d6b019ed97adef95902d5e402fb561f8a91f8b6a6383dc97d71579e8056e9f7aff8c530ac3bb1a47f4f69bfffa5e3433e6476bd4eca7db58b46d59f70dfc2d9d67f9247dcb1945dc8c29ebf28b5dea8642ff3da9a172513f6f65f9f0076a4fe5340163151896383b304cdd58147b21a4ed9a32bb2bc7793c9ec8353f91d98cd10f308c0e1887df9110473168db556098b39b318480a5ba31da45fcf0bbf8d28084d6fb3b2af1d72e9f134dcd21d343ab5afbd270f1f633230486073de7cf146afa09c51ce0a89dd841f18794d30d8bb50341fd47d73da390d25c4a7736d86cdfa22b3142774289809b32fbcc7955507bb9a340a5b45da00425da23dd8d565cfc7994704290e96a3e82125e195840aefaba3a9f6b7fd744ce5c13406e454e2b4fccdaff427aa06b27c753a076c163efe2a263171b3c5475b7bfaa383c9df080c22478a0d8b2fc4c4a9fc070f032d5dced9f9357d1621f5c5facd64f7ddfeaf5a4424522415f83bc193973848a06ee8339cbbf67c4442f52dab210e9480a8b59b3e69210e4b0d7cc6adbbd2b36e42cdf691ba88e34302e777766700e43abba4a035f8ae40068489ff01734535d57cf9519972db541d2611c20958d79fb55892c25d77b5c4ab6666b3daaf9d8c4ab4cb9fa4b3116d4eb6b22895fc2c512ab32b2800b6e07e05a23491a83367296c26d609b17de2c88abe1affea5dbe523ea9183a09fd6616b0ba8c74052e6be3d7da1f0a61478f9bd3d9c5e660727883150f02f1d002f87540c0448443b80b2df5467bf0208bad2513d56b0cf5489d8c1cc69c59c78ac55f64c5f15fdca0ab41d432341819a6967af62be6dbca689f7beed9b6c339ab435afdbcc22c910b6a07df6f9202fd40d90afbd8070d6dbaab3644699ee32cb224dc21ec84ab9af68e9d9535b6325e8f080e93a0b6fd4128a8107899ffbd6149707d51487a78794c45822ccfbf74ddce6db98482ec339a2ccd8d7375ef41322abc4d7f7f9ed65fe6200f49314402356b4eeff5000497b75ec0e935702a0c79a360b37db1dab933b5154622026c03e19e98e3b853ef0ca1fb910ffeff31635b294b066c1f6cf5378b4d8f6d88f4fea5dd18d4bfb39f6286b4d55b84973ed74cdc3260ce49debff60bc5fc6667aee2850ec60a2af8bb29d02e35241657b592211fd538e59ffddaaccf8ddca9cd579622374cec706aaf9bf45f861e4f62e5bf952c6abebee61ca8c5ce1e0bc667591f32c39df0e6d90b60f57d43b60efd13a16a37b59860067443ac482cb3a57fd7da90e902b7390236f2d7b4dd68844e2c53a2cf94094d925a52b7e3bfecb71fa34a05b7ee60dd43dd2bb922500f7d3c3a32e6fc40160413d63e3ccd1f6ab67aca281d86e8bb75604a3365981a9af20b82797304bdc01084ca4f890fb1190c49ef1871f32e86c54f0c7ec54b49178cac5b43bff7053bc373b4a5e235037b57547141f4eb88e227a0401c44e4cd56e95b45cdb42e5f255bb647544cebad5b2a7f0332a0376e5f7826cb20e64469a733f8a196e7b340ac29ba9aef92e0cc0634948d7946e5647b561e4b9799d8d3cc0033f04622b588b71ff590d4b6139fda38279ae41f53930cf710806e86eed854282bf243fe02ee157bbdcfe86fdc1e9be238f82c690fcecf9683c7074c57341d7e1ce8f21c6e6b09ae3cb654d000206d5b5b7b2519fc0b002658a2867ebd533e1ab139410e27f6e443af1df116b326f3d40c25eed5ef9313bdf18d04418dced1b48a2027c6a19640b6c764509725e3d9bcb5538540f511b5c88b4502037fea5a3cd5822a06e80e45ac5d3336f7539f0b8e03151d17b786d0c1bd0adafd41b0ca8877090725a79f02a66564620ddf8764dba32dacc7049bbcf2bf5902db85002402a8277692e6ffd32d6f6fdee1108cdcd336092ce1dd9daec5bf6b6170d8e5116dd37e9cbe52242af351247fc171aa8696a608f3b5730a5ae905fe00bc4f9e9c2f019be13e68dd27f695bbfb94e0e7f2fc459cb3f5e0303416d98c3349d91f720205e43ef412ddacd10b65097135420b63ad8054b48a2eb0b76dbb65d06814a271693355d2a06e55450928ac59fbb8b2855e125a64ad03101f28bdd2fb12e10ceda9feaeeaa52663e10d2e664bbb20aadd74a1d5e054162488dec6dc4373d9ac7ac3fd66443cd1cde2b805578fc287c7dfa83c67617b5609934bf2804ee3b8e1e4f56c048fc72c9e228ab687c4837e66080ec880e5dac3df55c2d8938b3367eb7f184c4346008a96e8c2a67e9460b3a1905c263aeb5b8db44bf2a082d771f4f0d320bbbf08226c496248ab834f4789b90d16d6d6f88bc3fdcda84ca23c7397a879c23b5b767b47e811304cd2df375094ca7fa61dfcc8947cbc22c973167e4cf6f5d6a853c59b769dbde3ff845b8c97fb38d6f15d69cc6edbb45540d04911aaf2b95df83491b03e61d6c5bad8fa2f4617924614944cd8d092d18d4039444b644834be9087719d7181080d59f1a54bb8e1f902d626ab6d648d8bcd22c273011156426972bd195847bbcfa9bb00124822325c18da5e352ad5a3d8db7fc0ebfd7f37138be2cc12b1ca15d2f075d975e9b26523876cdc0795480686cadd1ffee8c64496a8e02384f4a5bb012734ef37d4c59cc4772949876b093ff7dd6ed5d181858d238a79e8c19b5b44ff25d0c5c36e154885db0eecb7269ff25e89be5b50035662f1816a7b066b757be0ac3eb3b4d616ce11d6f2f0f8b010613c0392eb942051844275a024bf468ad26d096610eaa072ef64a44098891df0660c2ed1a608618710821b1575046ca29ac5c75705803ea8db3dd6495a5c2ad39222ae3a6dd6107cc9cf8a81d9019df40705227283a9df69a979b805a20986beff6723760d49f6539b73a515804ed49017c228a0dffa647f966b36b31a8e5a8ab5d640ac4cad1e6eb5776c36bd40443ed226afcc7c50ee23d86702538fee655c0bf69b48640fe555d854f244ff72bc56eb0ca24ccfd46813fe9e90f829fc170981875c3b7042217de239ac58ac57984f4b8828801bfdd35534f5edc1c562050454e33d6d29943cff7c9f2808056d646258187794afc2f11a7e701b4a5fc7f1270bfcf59344213cf17491756ad142b886720415e2268783e8ceb1e47edb8f6ae20319c9031119aa1c8cdaddc3573ee459bcea5c41088cb0203d7c60e54a36064ec6e757ec89b3917bf1883029ba067fb9d842c500d2bc2fb2ed4b64b7c5aa3176bceb00a9081bff10c2fefe6843becd609c321d0f8ef3c26f06ce32748cd7aa9227ac65a355ab9080e0d53f78d29fecffc5041fa64ecad54c06502eb8f0e7e1eed557ab3c9604535b62f6b0a683d9cbe2814fb6de2f12d531b52652764e0cc4a6811d6403eb43dc32306ae895170c494dee33b4be9b9ee7b29dad3431ad42db468958aef5f372bdac0f15d4cdddb14ba1fea9d7657fc344cd879afeb71844dba2ab9ff69abe55fc1f8a30ebaf3421e4eaad5389152578f90c193b1bb29876d948cf2092b5a8f0bbd4d0359e64be3b0d00b4a89c035ad2ce114beab01d0f573fdc254695c7d3fa5a8c835b10953216d820c7684ddc5258c85f5351e00f15650548c4cdf192100b769169421bd70cf918a20154a771bff5384daac43ad819f3da8ae27253fec9751764b77ec0e7a04d1b975132e4361deb18c06981ad6eb58a82eeab59e432ebe792ff56ed039f8914ba678eb30a8acde70627bbb36dd3891c4409f8fbb5359cdd974440fa6b15d3dfb72059e49e5a8619ade70c3046fb53f875aabcc7023795980ee078e7aa3299696ec51ae408e565da771c9eeaa8c53cffe07fd04cdc4f2e191bf8cea655267aef70aae172b86fa8eff2a9c4f5c9d9ca285a325010112b28e1eeda893a58a367920775d4abbbead18670387c05e97fd096e28c8f475087d78bd0493594dc5af131f26bc5042a1e2c219daf355ec0a38f17cd7cde9f80e242b9fd83dbba77deb0bab89e900b9be0a360b2cd4b8b121e452642682f186cf11b757c62373274dc4aa78d970b3a5bb1182098b80428fd614bbf3b1288c08041e74f9c1548b716ade180d3add50f193f6c013488f735956675802f30dbd302e627f91543d110bc2b5c7d5da94cfeb6d297a497995d0b5f3ee47e942b824d0a31d6371787f60c289487689788006cb6d027121199e0994bf8bdcdedb8d8e711b74b138846ce093018e490d71d02a7cdf8e187a740dc8fe465aff03a537564537e1428d6a58842d2ee17e295fa9caade6ba0717223475c665f0632d7d090ecb7e4f82a6db0db760dfaaccc84679b30fc32cc2505f0447b5179b32509c6c35823a1989d871122f2ffd12d0fcb6521dd57cf8d6faf88efad735a43ffadfc3faa18f5b5c3b57b2716adcdb24d9a69666866480beb1536e5e6c9b313e825e18209b62fb017ff3ea5be488b6eab114d12903d19046ad2a2830d4c5ca22d1d5b4472a70bc21f9ef898bd69bd5a75c1aca8c9ad52af1a710362a2833931c0138cc1385206b85d486e97463396e55e1776214746cdfc2c6b6a5a3390a016db903b772efc6ec9688e6d2d83387ecd30581eeba6f1cf52657cf9147a52c7750b3f4e86a72747c012e1d6acc9eb448b87f7a0d9706877f82bed0e76605a8bc64c141d8fe217c1969861616d10dac7f736513a6ecb7d2853fa7d206e507b4c34ab4ddcff8687c29a9a02229d37e09374b6ff89f294f3ede888d8e4e99cd16e602b26ff2575898ddd51a7ffe90d3bd60246d3dc25b9f0d28c72bfc53a5607dfe754ced1662b57b3201c433b264de94d9209a4c9d246323531f59299768206e757be6da55597ac7db29d9bc2cce21aa9d84a4aab53ebcb40ed8dbf6b46647c0e64db66f5301f4459631fbf2bef40c7b6b36cc56c8bcf079531c86de638e3ebaf906c73cb5d73b0d99f0e98d9c76fb871d0fd232cb8ed34f1d9c311857ff35d64b07c9640003594e499c76dd086ed6000cfc64cc158c195a18170c496592cd4718c4b3cf04e5b01fee8bf4d25ceb4c4041f560f0da14b99ed7c40c0c22722c0d5c0431a8dc17d5bfb23cf27645781ebd5d4cf64c2f800885144850e69f3bc40414cb1d404fd86231ce5995bdbd41696140ed1c46782a7893070fea4c66efc3201b97b95082e7fef5504a9ab34ad1b993a127d830d4c1f057550ee01b85e921fc15996fdff20f80690d4b1b97e1072a702ddaf0aa6aa27f584a8ca02c46b393b297db1e8a3f8ef53551310e2b7ac5092192866c598ac4fee8304000734cb048ba6b63b51ec017606705ad1c3789da191124aff0f5a6f3c22be9e1306b9462960f0271e12cb09bde69685793794f4a794a049ffa1c3ca50ba51cba14fa8a702889a2f5168a77b26e4ca84f82ca17da3bd00f75767ccf5b626d51fc70fa2c18531becbf2d2a8a11d106a921b58e3a9a23e20597a4e5cc1fd51b030a192fd4e1060330000000000000001bc465101206bff73890a12aa10f1b0fa36abd2a3aae6113a344a93abd55d2a9ff29a794f0692ae8e67b933321613735411b26ba81dbb7947dc319ed3d97093cc135fc8839998d85797a65a683a3812769df7a8ba069e39a20c261b9070c04c20df6ba04f2c6924c640003f4791b1367ef2e3d07c2e861093cb90dfedc4bbbd9d13207c6b2a22703563bff9e06b4a37faa0ec08ccad4d299cdbe0064eefdabce2878ed5e02e6d7bd23bc015206865065845d3ab1ef19c7f7c51916c254f7b6693395b27c3f7f708bd5a8bf7b9a45d48949bc78725f876361f3925b5a940e0a04fd487e65c92e0d7e17b8e0d3753a6b6878eeccd2f4e517983349613f44fe9e9d927a43799f713ec814a46e46cec2598eea6dbf5347c3af7a4520712bc93ca99ff14989237694bfd96493256fa378a10feabdd6037c89c9b50f64b145c66e479fac1aa9fc3dacb3f8709eb244fd3a3f66cff7a311ce99ef4b01f32d3a1f55840181ba870d21bbe6b7ba47123cd662c4a49a070c5890f9df3cab84c91a079638b6ce3c71804785a32b0f147cc75ee37efcf8971485c20f3d17a13fa14903d73783235714b445c44b28237f5375d111f6053322a52ce78a6f76e333874472a6d9fcc506ce10dd1c66b22aa4984a8ec6415617e5ccd6871e8c6bfe92e35d4396f94f1bdfe6dc989161bf4c5ae21493f820ff1cf0d6f44268fa66de3f666aa59ed1094f218ffc9b250586c0ec490479a169552d13f16220bc40b9aab6e692969cac82f113bfb8e968f2359c814f24066df9f85becd5f0a37efe01b31ea5e1996409d477b4e1905660e0c5e2024c50027e14d89bacdfcd6c350b7052f927e0909f967863070cc1ca3b2cc256488f2ed52a824efa9bf62e72c40caca46919486d2fbb74c92fb4039479085425af3c7100b0c5ad8da44ccebd0daa32abd29bd3e3f2bc82e0942e814bb8db78903fd73a4d9a4a842eda40280bb5e13b6334f971a1c5a2d043f3cabbcbc53c92e345f659bb6c09121e2c6e75295f5390540bef7b66e77e24f53588faa87a840bc7f6b44663262a1afbf766e80428817cef72a3edd76dfe32ca4048171fef1dda8d6013639cadd99f2f9c8f5da7a0c1a7adbe5558c5c19d2914a2613c0b18d5749898e15f4ac20e956d7a20573880fb53c392c1112c71fccbb93b8aba268ca3ddc056a65fbd21b76cc681fd417606a914a8056c9bc0f812a5e4ae8b8df856219cf9e116b30965807694d3ae8eca5fc66cf9fafe9fb2c701ca7b3456b5ccf4843411eb8f9e4ef6eef8c7953ef341e653e410f7885e94b2a0fb3ff08fd5078ccb726a153e6feda9890b3dc287f717d8a9d4f3251150ecd44bdd0bdcf744a86083c7ed27560130e5c714b01ea2e360abe43d65d16ee6e61af9955db68c047799c3642d8e70ce88995f818f5c27ef8cba8f62bf691d71299282053b113ee2ba993fd39d0592251ea85529bc3a3c7f51f5e6339101e1889a7d1b4c26b90c57926e01497ef1e58bfd5b358689eb65b2deda4dd290c014a284f0976f517e67d4f56d640744a3c5887bdbe2734b90374a089863280ed89d101843404bfa62ae6721c3985d31438ad812f9112f01f902078d126b995d96d31e889d9eb2beb6a67427cc1ad5ee113ea308359adc383708730de870bf54b4ffb891090a545ce5e90df29d90d6da5d102d27d746c07dd8ef252171665153f2d19bc25239a03c4809d569d2b7fdc3599473651a9430ffe35f9d0105a6cd85a6543c1ba5d6bb3f8a37b3e7fe8ab10906162952e280f4d288254adf239577d63edc1a4c2ab35ca37d31ec5f1a2671d9c3c8f306c67554e2b8ae51ce5d498b448eb13b572795e543972255813d8ea7f1b27f278abee9e480bf0171a58f9849c5dfdd708fcb75819ff604f37a2a5a2735dae1d619550553d4ce93900d15abd01e5410a3ddb02818ef3075332d4d8e42128d48b0079b9b6ade380b244506e1dca04c8d836082a7bf9f12c1220df4b3c33bb7182b7668a0837364c847477f0e24bed9402d5df53059f1f8307087ab9d943017cf996014ef0533b386fea55de964455721f1dc28c73172468a7c6642dc16c730db6137f94c8af637f45d6d83f61e12da78e1e577136a08acffab81ccf575c5ae14aac2485119987042cf7666857cdf6d90857b755dfceceff53ed85453e60427e9925e24fcca2c5c1dc8725fc2d4b9449a1bfec1cc7df92a322f164ad844b96417e74f6215a966800afd0a84f217e57f9fb930e536e7a4d9e3598c73c23f1d64f400354ea5e0db8b70a52a6187af137d2eb33b1a33b1464a675254e4aef3ee9add35734f1d9ea71fe983acac9765b1b8a57ef70ec591c02b5cfaddff2b7f2a22420a3c85024e72d651edc2b2a69c7d31dd20b4dfbe30425070b9f0a9a7bc2e6a4fd74fd4baf9772e54adb268780a7cb5b20ee8be5de24475f0e5f826d46ae24e3b213edb0da3d0027f46f2185ba7cf49e57d642ba5dd25edc0ec980d8014cafe52feaf6d88bf4428eca1ccdea28e31e04182f14ed2fc75923e91ce8fd5a7a4219e23303fc1f4ea98bd66c780e0152205e1780176793ef79c955b5005c40de8b4db30d3127a6769c13a09c3bb45f07c57f6f952de854d8feeba460bbc7f8119b8be9055a364e94a992763752da2e96f3eba7044e06eafda61b2699f14422970fb2d0ffdf9f7a80342da4f3387dbee3e8cf04ae9b707056efcae5bb74fd41fe954f133f4512dc967e4a413c02fd29d01714a18a1571c2c3943e77595057cd34f6cc44fae447ec961c8dc49eefcd0b0edd15ab9261dbb01e089c7ee347694614a166a97081757e941ac15cd1230717e926e51a2b6c25f8f83bbb3c6c9f7732f727846436cc4ea8c69a9527a26a8ba32047632836229be69c228b6155d63f362272e4fdf2cdd2d3c7f2f80350565aabc9e36a931e30216108588aeb901c61213e7906cc756cd8f43ed05211d136b1cad9e50897348caaf11da9bc7c3f4237f667c6e9082db00edc2771a056a215ee2f097a082a30f804f448123b6f90675e20137a914cb1b368579da1053766f3fda7997f5fd39ccaede40f2fd822f71131de065d72b5d838d1fa61f23143be8d5bd53f0d135535e28e753ea1c77708bef3241dc6560ac472c79bb1214cc8e95ac57820ff65d1fcfbfab5851fdcd9c5d67c0ce90b759f310f6bc1dec29b84dfba8e585e12497a3e547bec768878c320a7943f8f0248efc8147ac6634edc71a67421af209a00acd72b019bc338322a0da97d967d6b24f740c3ccda4aaab02f8b075dd97b06e4c4a3c494cf4e392932076c3ef09511a79b0d0e59776691e5d1b57a88c96da0471ff68ace378ce851f2c41d0176ba409e5eddd9c443148ab9736569c0e8aa18118c88efe3a49d0efa99e0a9d94c53402c8f10f6c56bc196b3849df4c2e65ab99aea84e066f552175d489edf0555f3fdb9cf046e92d421418d3cda02a00901fa59e973a2ea221741b3248e1f10b3085884221a2df712de5f2577f536d7a259ff381c654ed56263cf6ff5ed4a1ca15f97c3e30e872081c42eb8a79985567ea8f4f836df01e76a37acafa25cbfa2ce3d14284b65c1f51688be65084d663e3a24b8dd3a6c53515ba8b9f209642c803fda70baab6d7466371bf4586043f917452b886f81b369c9e31b152cf903f547259a20610983df2811317ad7dddf401ff028c3e98a9d618430411d22be8d95010eb5c012402e994aa72f5b4610ce18262345794c0fb44c03fdcd2ecc91a6e6cbfa6b2c4ab4d20390f651afc9b66471b383bb1b88af0dd93f7ed7f5eaf9dc996f01c2b26bbc0fea1012ee8a62eaa5d5576abce0a37826c78de8ac52d7be02031dc1ea5202900af551eabcc542285ea6c22682f29aa28f61f1362a225adde9a0437d5db3b5331333cf78c17601df31efe0f03a9cc1a4adab7cecbb422d9abc727940f1d5ab2a20839ad720e25bc5b2fc847588faf3ba2cea32b95e568cff1774742e2c18d45b9b6ac662c45345706253e58608dddaa815506b9e9a9c9469e29a3a0e398322819ecf74cc1f5db0a194063b1b224594698c0b38d4e8ed5c3bb8ed21649851bf0abb5a460168967976df003afb973cbb54ba2e000c418b78794ed88e0cb9783e746a9f3888fd567fd129aaf637166428c64022b1553328aaa3fd524449b5880da0fd94b4253e36b21e3270d5352e8fd3eeace5d526d5354368df424857c79d8cd39a9a85a86a99a7bebfcf272e5221a41396a460360db863fd6790a80d93af42ceb18d8cf4f33d0e6a8de08dedab8f07799ee6c9cd3985ef918cc7d6d8f0ce059e6ea6368994a22d678616ba123261bf62e71734fc6155556f7d12be94bc0c75163d5bab7505fa4b0a8b7102b0a9793025c63876b49a9872ddeef73c93b240cd8ee9bb6289f811e681149904de41129d2c7c923ffa427eb06e6a47e5cf622af755862b4ea1415193557f4a5b7c5ded51278a561bb45166bd1ea537ac618812b58ba1e68486b49ea7fc6d838c0551db27cf524d49bcac2b3cbd1f55f4cdccf1e7b0138edf0e77fca405f3ff2d1584ed9d6f4747e510e0b8194fc2f5e37688b90f87141bed3005f011494a5b866a17c5670f900723eefc4cdeada9f0690a07938e5111a213373ce5623d63a4c18fc2f0fc7d1ae618f4be90ea71bd227eb4fac6d05f9cad4bd92990d15a594cd16ea1668d15c5896feb2a77335ed6eceb3b69624b84a83a613e384cefcae32cb8d8797a688a2216b30efd62c2534116d9c18dbb19adc03f4df77fbe6f23c8d0aa71afe2b21a4e17f6cca97d2874cdc8f57e8a024096091fdcc11f6c573c81f32cfa0838f4267b534501136f8e9603b504bbf1637fa81cb34f0aeaa1013c579558ee2308898eb64c13d7d9844306da84a0d1f43ec3d1d158273d3075f1da3aa2b8f5db1749db75408150296e6f63b3503e21b0420a81f8092d4e7058e6025a8a5b8cce3e131b97dded447cb5eb87cf4a4ef6cdc10577d162d58a442831b07aed80d32eace2c8a85eff1afe1a05c18424562d069b0c57b97402e015d6b8e5598906292c09942f267b6f511d808f99d18fcc2ac83f9a0c0c5e24e2cb2453244fef227fb11093bfd68b1dc4ac1675fe45ca1363ecfed65d6bc94b5bf7f22eca626bcb75a2390023c6c0e2840aa572d85b2337b8a7ca4c58216082a4cc32e68cc129f0fb5def3d3d0a438a6a4bf81bcb6aa8e7c5c3ff866bb68c2b86400682248411978f9f247a3433cca781d0c6bf8da4c70ff564003e454409333ef1bb61d391d0e45cccdb41a6f3afab7fe99483fdf2598b6c2d9749ce7666207bfee75d51737f824e141d7aad8d31d10f368d0850924c753793f0ba02a24f6c084a4c73d6d28250d5332e4d2bb27ccdb95c8844c05948686a73b80be8d8d59a3e094b2076b4b2deedd00cd23a064f335409a89181ea5f9cd0302fb13f88dfa726d5c7039ff9bfed447cefdf283e943925d5a4ba9a7ee65a9e973b4e3ee56ad9740cf9ebe2be0536388ef3a84e1e17ed19d36591d2311120e7c59da1016d9ceec87d05bb73b04f87ff0b9ec6354a731c265d4d3bf5181bed5e3b9dee218a4840c56c60cd5382b12d35d22c62aaae90af4ab149f244cb0e21dfeae0bc21c2675b48249d340e6cf9b5ab7f80b2d5dfe69ef7acffe115ca0dc71a074e30531d4c39bfb313962375ac09f9deafe95d0251e6547c2f5fb831081f365e49711ff1da308f207bc888b2e14edb99fc06d567fa35525dd4cbc9105f565f9d7cc6baf9146ab5da76af8c77a5a08f318360616009cc907193e83dbc308a6dcb012a09173b8c038c5969ebd9f8ec9cb22c512de6fae2c22528b37dee61077c600e7e5859e2a59bc56cb76990981e91b112677eebc7233bcdc987e61b00008304000b8f969850764ddde29e5dc2d3bc64f3a3922eef846b8f41ee172ac5eb375658be34d651310ed0ae5baf4dd2e08d4f0d720bb7142c5ebc5e31cba7a10c64dd661e7ab5c592bd7b0cad789e49c274f2a5e8c7b30d8350bb43072ed09cd4afb66f8275cd5607a55d061f4fcbd18c70f3d1cf6c4cfa38bfdf50bf07c3b92f8239279cfd4e106033000000000000000117b96b009063e2a5083ea37b42389829404aac49bf57a9747405f0cdba1e98f0834d04ef48edfc8b3f763dbd15e1925bcd2d60cf1a9ae9aa90c059518e13877adce18833be3008d36f9d66824d22ba3b6cbe88790ee5440503b5cd8eb5c352966d36c3ebd88d824fc5b7bf62fcf599ac28e76e360c5bab6eabafa1911a048bb1410e1a8b76bfa0dd11276eaa109d447a7ed85c66412ae17804797970732187d63a67edc48845019cffb5635e9058e3b7dc36520b1b4ee35e66732c141703bb66ba12540d8b0a8356af623e8f97578f4e36d9cb3e1e40075b31a92fc6d71906aaef00b3f06cdf6c09ec40cc21c075766518b00938d3dd90b48115bc7c297ecd915f314c4f3dd84addf321d157690f7a559d1d24cf06c47c11d962526399e87c8d3443581b2071da5fe1f210d84c62ec8d9006e0600bcbce85bed8aaa1418c79dab3c4d28eebb74ed30fb6cbe6abe7629411d814fcaf11ac2ec84cdced52528da1093ebdb4350cb24adfb3b7bca2267418736d22cb1ae92fe064825ad749b13d559a4a82f3cba4a1ce56249554344cc13e076fa71cf1b6de2cc061d40ed56c41ece3755b3cf1354b67e065f49bd8bc76e074c5ad8f0d9858e92f70094eefdf83fa0f1232b8878707c8d17ac0ed347d66ee71fa7cba67079ee0b6d78e411bcc2909d90dfe7a2d8a0f5c4752a9fe9d66cff970730f6bf2cecc8e783d24bad48245cd1e28551eb5487066f80388571b5bc51b1a43d5c0ebdd4197c1cd1f07e0a711ce743ddc14840966f1e1c1207d7a6b659e8da9840a6c65696fcd38929b34844d36e69676ea79116748c23c8b355a079e0aad102dd09499fa7e208ef13f6264d6091f2df352b17bdbd182683239d0e300a91c1bbbf4cf9998fbab7c627804aa29ee9fc58affd736310beeb5cbd9ae29b327eac256f12818b948e45992cf89186baf3fd992717233d844fa7373bfcf8d0742a8a38fcd6b7258b47a2ab23281ff03c69ab68f336efa5e6a6ce7d9ca9a689333bb45662923b171777d7810c5d17fe8cb4f2cfd8ea0fde93e280dc17959e54a2580f88049c336959a3ee3d6fdd33283e906c3676ea1fc8c2645440f3ec6d753980d006f0481a3be84dbcb2de646024a2b9da93873957b486917bb07168da3b3d839ed11f0050a61ac3cff125382a439bb60ba5288babeb4a71aa957e2bbb3273c5e82a643f791e7ba2a8a93ed2736baee2e199cb5c8ab6e1053d730849809a412bc5dc1ded7e9d18b1ec2a33cdfed8c4b6c28a183155df8ff0ab0f1da0bbfce767f55e2525f731fdf05e88fbbc045c3ecfc23e3f6e1e7d57d619c88db95ac35652291d3b78fdc61e45f62ad226306f39a031c41d1ee3ed630f2cc52deada2b0d8dc5dc268f5dd0de062615ad28c879cbd1c94092c77d5cf92bc774a44901564433a1a6645ab31deb3902c8d459aaa18d4b6eba4e5c63a46751981bb8d3b51aff39e064c798cf4ebecb16dcfce8a5f3ceeaf91d7f62807417287c7de009eff2a85ba4a70972f40540e0739a443cd1552c47d2b780e3320d3c514321918dc33e4ecf596f7a344ae2241e044ff70d06e134e68219fd5ffcb7a12aef1f1f31766c301935ed667ee57376fa856b8fe3a499cdc7d33e6680e12ae8e6285bfc63a0325da883d30ffa0498e4fe2154ed843e1f180f2a058e6dcd0a1cdbd138c4eaf389636954454ffbe928805d0ae7a322c51a105641b3bda84d96cb3706f60ced4a2ef3a7cd63cc17c053d2ec9702bc7542d1e4db98db703604603f1dfe777b63454f4f239b15ec9458209c77b7e60c735f919ad9070fdf85891dd3296adb745305d3d215ea76e644a1dca72d3e1cbe199b0c11b18928beeaed410766c8f98e50bb44fc3b19b8d502c8ad61930b16004433cf0c322e9ab6c4faceb06e3ca25826883a67486ee1df89d0b8b2929ee351916b4aead312047d4320646a3ba368a3e886e1c6e7eada9087ef5a84e79b0ab0a3796f5ab0cf1217aa07d1ae1387c2dc6092cd02b6ab6b3c65ab761a8976e74b251128bff432d7ce6de00ccd711c30838978b0c4dc328502b5b42906d401f5b66fd175285ca7d59310084357a5915c43ba89c77971d6cb0db6354b2ecf14d03dd624073ef754149cdab8af01fb069394446b408b48b6a0bd2ec14562c73b4ee44fecec39b7ae26a8f8ab706706f030e2edf4a2235d11b798896b738e2441a6693d9f1dd8af39d9f95731fbc7aecfd056129a9bcfdeecb46581fac573c844e98e4f8ce74fe60da2cd8f19c48efb6da2521972a840bf38bd9ad57be2e68cb60adc0e859ae5505e230c82e51eecdf38e80ed5e73d1d9f7838aa14355f22bab25a91a3c60b3391b71d8995e3a082a37bdc5eca47790828bd269135715794bf34d4c05b686cc0c52d6e62fb274f564c69b0801906e07f7499a0865fd27db999f574292b6edffbf1abf678ef0ff2ca869b5c4f05c0d5a832362e9c7af91d5f4d81923c7232d1c2e4693433dac36b4ceb8d300ddf5493867965ca4e8feb8a5961952a76dc750d9ce733d318f43ffe02431493fe7e86a22a3d1a71cd86e8afc6f1e64be83ed0759249caf8cc3e05adb513760a1427abc90043b9662030530822662baca4febbbf034564168dba584f299acf7537ee4d8d6964ce5d96e858750cfb0c4eee87d483eb79237a8bab53a1081911cdd82d14979f26d33bfdcf5e19ef216dea8503c89f07dc579b572ec3a444017a787996325392f876417e83b2eebc525eca41277f2baf98255d1b70e75e3107057cfc66215e152889fc50aa31bf133d1fec7d24f215624e6357cf0ec0c9816393181693a14f9cca48504946f68277066685073cb9af8fea22fbee2a5570708ce2e6a44fcd68ad2222a8e03af33eb70a0f6fe871af8c5b688eecf290f757560abb6338bbb0f849ab39d0f26258ea17385d7ae59bafea0fc799d36197ac5aaed9b485e10c8bdf6f931ff7814b049731c24c16d508442d1c4fcc3d01d8dfb557e9c0e2afea87315969fb5ea07c4697d7368db25b2b2192461109a5b2fca89b7b63709fbb3327d2b00fc6d6f9e16c83b202dc8bd26c5bd1ca1693cb1256950ca7beacd2c61382e103d9510d1a5cddc7b285603920ab9fa6bcd77cab566b30d7454fe50aead926c9e5edb22bd9ae79f0bcef80db0b6fe1e43a00be0ca23b89d5cffa113200b4b4f3041cae7d9623f3b1e5a46f04fc95d33d6e931219dda120d72456ac753b0c02fc509ebb4f020363fe7901b82a01a14c9ee2c09f1342c031f0e74a7a939715bc725fc5896c5754e0a618bd50cdef2ffc2e65ce3f78da4fb6a7e3d07166d8bd14b42cd44136826b87732e837bb4ac5af23988be02961d78361fd07ffee65d19157cf9cfa659286d6415db5c05f0d8065f8e501fda8cb3a6a596cd36086d30fb6f7d6df8efd6e3160549b40194bea8a631d2906e713d9ee5798eac674887466b2b1fb63b8097d672ce62478b0e4c7f5a8ba85c0227b3d51ea0763116cc46e47aba225bc6d3e0ed7a66b6a98284c6fc24176998d467c3e44b2cc23984abd21aebf5520e4f7807ed03969b8827e39974da72c3993b09284102170836250efaea4b465c45ad3cfdf6c823cecbabe66697d0b8dd8f47863d6d9c9d64dc8d6495975479f330dd31ebc9bf1e9e05205e553287c6b378b8424caf9e4caad026f09556b1a9fa91d5740daae8a45de26a0ae836cfa6c05da141981e7ff35c1448bd0799381c88a3e8d6148c9bfb4d26aee441cd330800655ab078177389a40b7ee801848b463724094d09642332d5aaa234029155f5958284cbd3dd4a3cf8da286e49c896df6d8728e6e9ad6180051e75907c3450ec373021c4d784f3f62b49cfb1e47624ec69367260d97e7eea9e104ef76ccd12ccac80495c0af72b43789f4b814aad8b762460ccbfe0aa84591624dd9dbdce72c6e5d72f943305759c0baabc294f3c062a86d348bdb33bf5d6c40616ffa85cbf59930d505c6a0cb36b8af2b151dc8839d1957cb8de88de20ea6c3e78d12b88867efada2fb6f12df80638bff1e1594722fb0d667b58f4c6c17b7aafad4b59b625bbeeceaf29935d96126a8427642677dd5d0f6cec3c3de7b2d41f6425ed71d7d66f61c4b4ef4ab1c5c5f5695e0abad6f65266bb3a705e7b859d0d0ab6ae20131fdde6997e34f5a7b506b4bdbad724373faa4bbaab3a0e7ef51f62107cdb3b78f60693d0cd6554d8b45a059050d70875e6553cf6723e9a00506d26c66d5ff12cbd76fdf7290f28517f57dd3b18b1986287f035ef2847df56a2a31af96dbd3d97721e4a167cebbb665ae998fcc27da46f8deb5454158747e6fcd9a6defd894a4c1b7053258cc88c58b17a8d8091c07f8fd0cd9a90cc0a71b3ea4cfb75ccaeecb9761dfbccd58326af050b40b0af57d2f7f47dc84dcb41a9447c1a04cd887acd214397a80f15a9ebbb81cdf59bf06d90f4b2d60dfc7c76b63b140d5fc60b5a5d9bedea3f817a087c354db619c1e333347b7929a9ff95dfd1b891991878cb86f6fdf78781c697ea572bff23572f3fd22014955895cb52089215e79cee7df038a170e4bccd4027b049a9220b24da95cc31db9d59ea82ee0152998214de40b3143aab42519226ef1ea02536bbdbac8beccc7d7090ee5f62875925740ab4af5f5ae3da9e30ac90fc429a161b44cb316a6f12004b051fe7d308b5abb3eede0409054e641f468b40a93fff94c0e3d704f7d9666c337e7d61ae6255441b15be8c3fd68e461b0d15a72ad231aad3d65b7728f85815bf3f29fe1f382a38c675c4356cb4c848b6121d3c8c7a15747df094ae9e4a7396c5fa85926c6285258a277e3eccaf98aafb9e0c2b8940f1b49081353fafc155a888e4feb2e30d0939a50bb80d1188dd262d2c191175b7135cc13bffcca1d1253218156c5deaf31842e6a34f6b2736707a2ccfc181f31b19533b5f21c4987912ae7b8f76f253bef706a4c0d537f890cfb7578a24fac431853d5f6f046a3924c45866edfa5a4f33802ec1c57213abad755afc7ceeb69347d9381ed05641bec9eb7527774fd2053dffcef58f7e654da291ee2ff745d43fe096dd1ece2f07468e64e1fc188ae1ed046457e14dad92feafd1bad54e0f87bd59bcb6cd6160431067b9cde6d4f132a41e862fa925dd734f7241280e0efa59a6317df389b2ba770233bc2e506f6c5512a944ccca53a58985f3c4744f4bf8a3c52f7d7f4a5ebc077a0b0d48317f2c2c001c2dd3b206d261e3a1118a6d44945e34284a5e868bd7e4a1292c84b95dc91568fe96d3e09d155f81dbbd82e6a0eab6578c2068a30b048a945f6ae00a297d88e3b3af4ce60c133e101384b806d0b6b707c00d1284e9bd6ecabb10729a04adeacd145eb9a5b0431abb3594d1cb07edb8b7614b2a2e36361325f99df79375159c787f61a44fdbead5d48cc200b96ef0e8af4da218479885504ec9420f232747b9dde8f593c9096f3a407b22e22c9168294c2e0322519340084f51ab7aea1dd5027e3cb7cab7061b3fd573c61edccbd23775525cb17da9fe6c4de4665a62d83fa843eb567599210bc9e98d607572008de038770b3dab2533610a74a0358821ed0d0b6a05f6a2bcbe6c7fa07237bbf64ae66af64f78d44ee0743fe7a0b14b89ce1a9e5570d8c3b1ded22bdcd543b3a96b0e13ecd3e0663fbc7defeabf10bb224632cd219b963b7ce3ea42ae422d2318bfb4c4fb86bb640fc8c1f6ec2362560e20cb3ddbad2690596ab889ceec5c1609899fe6bcf7f83b3894674e52250bfd87bcf45c4c96080594b595881a1021bddfdddb1d04619a283bc95f5c5d476fcc5beafe2b78acb6713a0cb03fd2e12521e6c2d4f8d5d748feecfcf038bebb9aff49f753fa9db8c61cb52d25626f3864accf53be13e274af8';
 const walletUtxoList = [
@@ -21,18 +19,20 @@ const walletUtxoList = [
 ];
 const authorizationSignature = '30430220307607146431862975de982412170cc9b45f3cfa2abbd351c634840b785f2785021f3c478b9de2cb55801fb924c5c219ebe71d58f8d09f6df32d8e17c105ded6a3';
 
-let lastConnectedApp = LedgerLib.ApplicationType.Empty;
+let lastConnectedApp = ApplicationType.Empty;
 
 // This a very basic example
 // Ideally you should not run this code in main thread
 // but run it in a dedicated node.js process
 function getLedgerInfo() {
   console.log('getLedgerInfo call.');
-  const liquidLib = new LedgerLib.LedgerLiquidWrapper('liquidv1');
+  const liquidLib = new LedgerLiquidWrapper('liquidv1');
   return liquidLib.connect(10, '')
     .then(status => {
       if (status.success) {
+        console.log('connect success. ', status);
         return liquidLib.getApplicationInfo().then(result => {
+          console.log('getApplicationInfo called. ', result);
           if (result.success) {
             lastConnectedApp = result.name;
           }
@@ -42,23 +42,23 @@ function getLedgerInfo() {
       console.log('connect fail. ', status);
       return status;
     })
-    .finally(() => {
-      return liquidLib.disconnect().then(() => {
-        console.log('getLedgerInfo disconnect.')
-      });
+    .catch((e) => {
+      console.log(e);
     });
 }
 
 function getXpubkeyInfo(path) {
   console.log(`getXpubkeyInfo call. path:${path}`);
   const netType =
-      (lastConnectedApp === LedgerLib.ApplicationType.LiquidHeadless) ?
-          LedgerLib.NetworkType.LiquidV1 : LedgerLib.NetworkType.Regtest;
-  const liquidLib = new LedgerLib.LedgerLiquidWrapper(netType, true);
+      (lastConnectedApp === ApplicationType.LiquidHeadless) ?
+          NetworkType.LiquidV1 : NetworkType.Regtest;
+  const liquidLib = new LedgerLiquidWrapper(netType, true);
   return liquidLib.connect(0, '')
     .then(status => {
       if (status.success) {
+        console.log('connect success. ', status);
         return liquidLib.getXpubKey(path).then(result => {
+          console.log('getXpubKey success. ', result);
           return result;
         });
       }
@@ -67,24 +67,21 @@ function getXpubkeyInfo(path) {
     })
     .catch(e => {
       console.warn(e);
-    })
-    .finally(() => {
-      return liquidLib.disconnect().then(() => {
-        console.log('getXpubkeyInfo disconnect.')
-      });
     });
 }
 
 function setAuthorizationKey(key) {
   console.log('setAuthorizationKey call. key:', key);
   const netType =
-      (lastConnectedApp === LedgerLib.ApplicationType.LiquidHeadless) ?
-          LedgerLib.NetworkType.LiquidV1 : LedgerLib.NetworkType.Regtest;
-  const liquidLib = new LedgerLib.LedgerLiquidWrapper(netType, true);
+      (lastConnectedApp === ApplicationType.LiquidHeadless) ?
+          NetworkType.LiquidV1 : NetworkType.Regtest;
+  const liquidLib = new LedgerLiquidWrapper(netType, true);
   return liquidLib.connect(0, '')
     .then(status => {
       if (status.success) {
+        console.log('connect success. ', status);
         return liquidLib.setupHeadlessAuthorization(key).then(result => {
+          console.log('setupHeadlessAuthorization success. ', result);
           return result;
         });
       }
@@ -93,24 +90,21 @@ function setAuthorizationKey(key) {
     })
     .catch(e => {
       console.warn(e);
-    })
-    .finally(() => {
-      return liquidLib.disconnect().then(() => {
-        console.log('setAuthorizationKey disconnect.')
-      });
     });
 }
 
 function hasAuthorizationKeyForTest() {
   console.log('hasAuthorizationKeyForTest call.');
   const netType =
-      (lastConnectedApp === LedgerLib.ApplicationType.LiquidHeadless) ?
-          LedgerLib.NetworkType.LiquidV1 : LedgerLib.NetworkType.Regtest;
-  const liquidLib = new LedgerLib.LedgerLiquidWrapper(netType, true);
+      (lastConnectedApp === ApplicationType.LiquidHeadless) ?
+          NetworkType.LiquidV1 : NetworkType.Regtest;
+  const liquidLib = new LedgerLiquidWrapper(netType, true);
   return liquidLib.connect(0, '')
     .then(status => {
       if (status.success) {
+        console.log('connect success. ', status);
         return liquidLib.getSignature(proposalTx, walletUtxoList, authorizationSignature).then(result => {
+          console.log('getSignature success. ', result);
           return {'success': result.success};
         });
       }
@@ -119,109 +113,211 @@ function hasAuthorizationKeyForTest() {
     })
     .catch(e => {
       console.warn(e);
-    })
-    .finally(() => {
-      return liquidLib.disconnect().then(() => {
-        console.log('hasAuthorizationKeyForTest disconnect.')
-      });
     });
 }
 
-// Keep a global reference of the window object, if you don't, the window will
-// be closed automatically when the JavaScript object is garbage collected.
-let mainWindow;
-
-function createWindow() {
-  // Create the browser window.
-  mainWindow = new BrowserWindow({
-     webPreferences: {
-      preload: path.join(__dirname, 'preload.js'),
-    }, width: 600, height: 540 });
-
-  // and load the index.html of the app.
-  mainWindow.loadURL(url.format({
-    pathname: path.join(__dirname, 'index.html'),
-    protocol: 'file:',
-    slashes: true
-  }))
-
-  LedgerLib.startUsbDetectMonitoring();
-
-  // Open the DevTools.
-  // mainWindow.webContents.openDevTools()
-
-  // Emitted when the window is closed.
-  mainWindow.on("closed", function() {
-    // Dereference the window object, usually you would store windows
-    // in an array if your app supports multi windows, this is the time
-    // when you should delete the corresponding element.
-    mainWindow = null;
-  });
-
-  // ~~~ BASIC LEDGER EXAMPLE ~~~
-
-  ipcMain.on("requestLedgerInfo", (event) => {
-    getLedgerInfo().then(result => {
-      console.log('ledgerInfo');
-      console.log(result);
-      mainWindow.webContents.send("ledgerInfo", result);
-    });
-  });
-
-  ipcMain.on("requestLedgerXpubkey", (event, path, asset) => {
-    getXpubkeyInfo(path).then(result => {
-      console.log('ledgerXpubkey');
-      console.log(result);
-      mainWindow.webContents.send("ledgerXpubkey", result, asset);
-    });
-  });
-
-  ipcMain.on("setupAuthorizationKey", (event, arg) => {
-    setAuthorizationKey(arg).then(result => {
-      console.log('setupAuthorizationKeyResponse');
-      console.log(result);
-      mainWindow.webContents.send("setupAuthorizationKeyResponse", result);
-    });
-  });
-
-  ipcMain.on("checkTestAuthorizationKey", (event) => {
-    hasAuthorizationKeyForTest().then(result => {
-      console.log('checkTestAuthorizationKeyResponse');
-      console.log(result);
-      mainWindow.webContents.send("checkTestAuthorizationKeyResponse", result);
-    });
-  });
-
-
-  // ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+function changeDisable(disabled, connectDisabled = undefined) {
+  const fieldNames = [
+    'lbtcXpubkey',
+    'jpysXpubkey',
+    'requestLbtcXpubkey',
+    'requestJpysXpubkey',
+    'lbtcBip32path',
+    'jpysBip32path',
+    'authPubkey',
+    'requestAuthkey',
+    'authkeyResult',
+    'requestCheckTestAuthkey',
+    'checkTestAuthkeyResult',
+    'connect',
+  ];
+  for (const name of fieldNames) {
+    const field = document.getElementById(name);
+    if (name === 'connect') {
+      if (typeof connectDisabled === 'boolean') {
+        field.disabled = connectDisabled;
+      } else {
+        field.disabled = disabled;
+      }
+    } else {
+      field.disabled = disabled;
+    }
+  }
 }
 
-// This method will be called when Electron has finished
-// initialization and is ready to create browser windows.
-// Some APIs can only be used after this event occurs.
-app.on("ready", createWindow);
+function checkDisconnect(arg) {
+  if (('disconnect' in arg) && (arg.disconnect ===  true)) {
+    changeDisable(true, false);
+    document.getElementById('app-name').innerHTML = `${appNameTitle}: -`;
+    document.getElementById('connectResponse').value = arg.errorMessage;
+    document.getElementById('connect').disabled = false;
+  } else {
+    changeDisable(false, true);
+  }
+}
 
-// Quit when all windows are closed.
-app.on("window-all-closed", function() {
-  // On macOS it is common for applications and their menu bar
-  // to stay active until the user quits explicitly with Cmd + Q
-  LedgerLib.finishUsbDetectMonitoring();
-  if (process.platform !== "darwin") {
-    app.quit();
+function getXpubkey(asset){
+  const bip32path = document.getElementById(asset + 'Bip32path');
+  const xpubkeyField = document.getElementById(asset + 'Xpubkey');
+  if (bip32path.value.length === 0) {
+    return;
+  }
+  const pathArr = bip32path.value.split('/');
+  let num = 0;
+  for (const path of pathArr) {
+    if (!path) {
+      xpubkeyField.value = 'invalid bip32 path.';
+      return;
+    }
+    let checkLen = path.length;
+    if (path === 'm' || path === 'M') {
+      continue;
+    } else if (path.charAt(path.length - 1) === '\'' ||
+        path.charAt(path.length - 1) === 'h' ||
+        path.charAt(path.length - 1) === 'H') {
+      // hardened
+      checkLen = path.length - 1;
+    }
+    const checkPath = path.substring(0, checkLen);
+    if (isNaN(checkPath)) {
+      xpubkeyField.value = 'invalid bip32 path.';
+      return;
+    }
+    ++num;
+  }
+  if (num === 0) {
+    xpubkeyField.value = 'invalid bip32 path.';
+    return;
+  }
+  changeDisable(true);
+
+  getXpubkeyInfo(bip32path.value).then((arg) => {
+    if (arg.success) {
+      xpubkeyField.value = arg.xpubKey;
+      changeDisable(false, true);
+    } else {
+      xpubkeyField.value = arg.errorMessage;
+      checkDisconnect(arg);
+    }
+  });
+}
+
+function setAuthKey(){
+  const authPubkey = document.getElementById('authPubkey');
+  const authkeyField = document.getElementById("authkeyResult");
+  if (authPubkey.value.length !== 130) {
+    authkeyField.value = 'invalid authorization pubkey. pubkey is uncompress only.';
+    return;
+  }
+  if (!/^[0-9a-f]{130}$/i.test(authPubkey.value)) {
+    authkeyField.value = 'invalid authorization pubkey. character is hex only.';
+    return;
+  }
+  changeDisable(true);
+
+  setAuthorizationKey(authPubkey.value).then((arg) => {
+    if (arg.success) {
+      authkeyField.value = 'success';
+      changeDisable(false, true);
+    } else if (arg.errorMessage === 'CONDITIONS_OF_USE_NOT_SATISFIED') {
+      authkeyField.value = 'success (already registed)';
+      changeDisable(false, true);
+    } else {
+      authkeyField.value = arg.errorMessage;
+      checkDisconnect(arg);
+    }
+  });
+}
+
+function checkTestAuthkey(){
+  const authkeyField = document.getElementById("checkTestAuthkeyResult");
+  authkeyField.value = 'during check...';
+  changeDisable(true);
+
+  hasAuthorizationKeyForTest().then((arg) => {
+    if (arg.success) {
+      authkeyField.value = 'authorization pubkey is testing key';
+      changeDisable(false, true);
+    } else {
+      authkeyField.value = 'other key (not testing key), or other error';
+      checkDisconnect(arg);
+    }
+  });
+}
+
+function connectLedger(){
+  changeDisable(true);
+  document.getElementById('connectResponse').value = 'check connection...';
+  getLedgerInfo().then((arg) => {
+    if (arg.success) {
+      const ver = `v${arg.version.major}.${arg.version.minor}.${arg.version.patch}`
+      document.getElementById('app-name').innerHTML = `${appNameTitle}: ${arg.name} (${ver})`;
+      document.getElementById('connectResponse').value = 'connected';
+      if (arg.name === 'Liquid Hless') {
+        document.getElementById('authPubkey').value = productAuthPubkey;
+      } else {
+        document.getElementById('authPubkey').value = testingAuthPubkey;
+      }
+      changeDisable(false, true);
+    } else {
+      changeDisable(true, false);
+      document.getElementById('app-name').innerHTML = 'name: -';
+      document.getElementById('connectResponse').value = arg.errorMessage;
+      document.getElementById('connect').disabled = false;
+    }
+  });
+}
+
+document.getElementById('connect').addEventListener('click', () => {
+  try {
+    connectLedger();
+  } catch (e) {
+    console.log(e);
+    changeDisable(true, false);
   }
 });
 
-app.on("before-quit", function() {
-  LedgerLib.finishUsbDetectMonitoring();
-});
-
-app.on("activate", function() {
-  // On macOS it's common to re-create a window in the app when the
-  // dock icon is clicked and there are no other windows open.
-  if (mainWindow === null) {
-    createWindow();
+document.getElementById('requestLbtcXpubkey').addEventListener('click', () => {
+  try {
+    getXpubkey('lbtc');
+  } catch (e) {
+    console.log(e);
+    changeDisable(true, false);
   }
 });
 
-// In this file you can include the rest of your app's specific main process
-// code. You can also put them in separate files and require them here.
+document.getElementById('requestJpysXpubkey').addEventListener('click', () => {
+  try {
+    getXpubkey('jpys');
+  } catch (e) {
+    console.log(e);
+    changeDisable(true, false);
+  }
+});
+
+document.getElementById('requestAuthkey').addEventListener('click', () => {
+  try {
+    setAuthKey();
+  } catch (e) {
+    console.log(e);
+    changeDisable(true, false);
+  }
+});
+
+document.getElementById('requestCheckTestAuthkey').addEventListener('click', () => {
+  try {
+    checkTestAuthkey();
+  } catch (e) {
+    console.log(e);
+    changeDisable(true, false);
+  }
+});
+
+try {
+  // first execute
+  document.getElementById('authPubkey').value = testingAuthPubkey;
+  changeDisable(true, false);
+  document.getElementById('connectResponse').value = 'wait connection...';
+} catch (e) {
+  console.log(e);
+}
