+++ webpack.config.js
@@ -1,0 +1,67 @@
+const path = require('path');
+const webpack = require('webpack');
+const CopyFilePlugin = require("copy-webpack-plugin");
+const WriteFilePlugin = require("write-file-webpack-plugin");
+
+module.exports = {
+  entry: './src/index.js',
+  output: {
+    filename: 'main.js',
+    path: path.resolve(__dirname, 'dist'),
+  },
+  mode: 'development',
+  resolve: {
+    alias: {
+      assert: "assert",
+      buffer: "buffer",
+      console: "console-browserify",
+      constants: "constants-browserify",
+      crypto: "crypto-browserify",
+      domain: "domain-browser",
+      events: "events",
+      http: "stream-http",
+      https: "https-browserify",
+      os: "os-browserify/browser",
+      path: "path-browserify",
+      punycode: "punycode",
+      process: "process/browser",
+      querystring: "querystring-es3",
+      stream: "stream-browserify",
+      _stream_duplex: "readable-stream/duplex",
+      _stream_passthrough: "readable-stream/passthrough",
+      _stream_readable: "readable-stream/readable",
+      _stream_transform: "readable-stream/transform",
+      _stream_writable: "readable-stream/writable",
+      string_decoder: "string_decoder",
+      sys: "util",
+      timers: "timers-browserify",
+      tty: "tty-browserify",
+      url: "url",
+      util: "util",
+      vm: "vm-browserify",
+      zlib: "browserify-zlib"
+    },
+    fallback: {
+      child_process: false,
+      fs: false,
+      crypto: false,
+      net: false,
+      tls: false
+    }
+  },
+  plugins: [
+    new webpack.ProvidePlugin({
+      Buffer: ['buffer', 'Buffer'],
+    }),
+    new CopyFilePlugin({
+        patterns: [
+          {
+            from: "index.html",
+            to: ".",
+            context: `${__dirname}/src`
+          },
+        ],
+      }),
+    new WriteFilePlugin()
+  ]
+};
